// client/college-sports-plus/src/app/shared/models/roster/soccer-roster.model.ts
import { LeagueAthleteModel } from '../league-athlete.model';

export class SoccerRosterModel {
  Bench1: LeagueAthleteModel | undefined;
  Bench2: LeagueAthleteModel | undefined;
  Bench3: LeagueAthleteModel | undefined;
  Bench4: LeagueAthleteModel | undefined;
  Bench5: LeagueAthleteModel | undefined;
  Bench6: LeagueAthleteModel | undefined;
}