// client/college-sports-plus/src/app/shared/services/dl/models/roster/football-roster-dl.model.ts
export class FootballRosterDLModel {
  FirstTeamQB: string;
  FirstTeamRB1: string;
  FirstTeamRB2: string;
  FirstTeamWR1: string;
  FirstTeamWR2: string;
  FirstTeamTE: string;
  FirstTeamFLEX: string;
  FirstTeamDST: string;
  FirstTeamK: string;
  SecondTeamQB: string;
  SecondTeamRB1: string;
  SecondTeamRB2: string;
  SecondTeamWR1: string;
  SecondTeamWR2: string;
  SecondTeamTE: string;
  SecondTeamFLEX: string;
  SecondTeamDST: string;
  SecondTeamK: string;
  Bench1: string;
  Bench2: string;
  Bench3: string;
  Bench4: string;
  Bench5: string;
  Bench6: string;
  IR: string;

  constructor() {
    this.FirstTeamQB = '';
    this.FirstTeamRB1 = '';
    this.FirstTeamRB2 = '';
    this.FirstTeamWR1 = '';
    this.FirstTeamWR2 = '';
    this.FirstTeamTE = '';
    this.FirstTeamFLEX = '';
    this.FirstTeamDST = '';
    this.FirstTeamK = '';
    this.SecondTeamQB = '';
    this.SecondTeamRB1 = '';
    this.SecondTeamRB2 = '';
    this.SecondTeamWR1 = '';
    this.SecondTeamWR2 = '';
    this.SecondTeamTE = '';
    this.SecondTeamFLEX = '';
    this.SecondTeamDST = '';
    this.SecondTeamK = '';
    this.Bench1 = '';
    this.Bench2 = '';
    this.Bench3 = '';
    this.Bench4 = '';
    this.Bench5 = '';
    this.Bench6 = '';
    this.IR = '';
  }
}